Refactor OpenAPI parameter and schema handling
Refactored and reorganized code in the `TinyHelpers.AspNetCore` namespace:
- Updated `OpenApiParametersOperationFilter` to include `Swashbuckle.AspNetCore.SwaggerGen` and removed unnecessary null checks.
- Moved `OpenApiOperationOptions` to its own file and simplified its initialization.
- Moved `OpenApiSchemaHelper` to its own file, retaining its schema creation methods.
- Added conditional compilation directives (`#if NET9_0_OR_GREATER`) for .NET 9.0+ compatibility.
- Updated `OpenApiParametersOperationTransformer` similarly to `OpenApiParametersOperationFilter`.

These changes improve code organization, readability, and maintainability.
