+++ src/TinyHelpers.AspNetCore.Swashbuckle/OpenApiOperationOptions.cs
@@ -1,0 +1,12 @@
+﻿using Microsoft.OpenApi.Models;
+
+namespace TinyHelpers.AspNetCore.Swagger;
+
+public class OpenApiOperationOptions
+{
+    internal OpenApiOperationOptions()
+    {
+    }
+
+    public IList<OpenApiParameter> Parameters { get; } = [];
+}
